Removing highlight changes from this branch
